+++ packages/storybook/src/web-components/alert.stories.tsx
@@ -4,12 +4,12 @@ import { Heading, Paragraph } from '@rijkshuisstijl-community/components-react';
 import { AlertWebComponent } from '@rijkshuisstijl-community/web-components';
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { mergeMarkdown } from '../../helpers/merge-markdown';
-import readme from '../community/alert.md?raw';
+import readme from '../components-react/alert.md?raw';
 
 AlertWebComponent.define();
 
 const meta = {
-  title: 'Web Components/Alert',
+  title: 'Alert',
   id: 'rhc-alert-web',
   component: 'rhc-alert',
   argTypes: {