+++ packages/storybook/src/web-components/form-field-text-input.stories.tsx
@@ -2,7 +2,7 @@ import { FormFieldTextInputWebComponent } from '@rijkshuisstijl-community/web-co
 import type { Meta, StoryObj } from '@storybook/react';
 import { createElement } from 'react';
 import { mergeMarkdown } from '../../helpers/merge-markdown';
-import readme from '../community/form-field-text-input.md?raw';
+import readme from '../components-react/form-field-text-input.md?raw';
 
 FormFieldTextInputWebComponent.define();
 
@@ -33,7 +33,7 @@ const TextInputWrapper = ({
 };
 
 const meta = {
-  title: 'Web Components/Form Field/Form Field Text Input',
+  title: 'Form Field/Form Field Text Input',
   id: 'rhc-form-field-text-input-web',
   component: TextInputWrapper,
   argTypes: {