+++ packages/components-twig/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @rijkshuisstijl-community/components-twig
 
+## 2.0.0
+
+### Major Changes
+
+- 8e508f9: Removed the following tokens:
+  Used but unnecessary:
+  --utrecht-alert-heading-line-height
+  --utrecht-alert-message-line-height
+
+  Unused:
+  --utrecht-alert-heading-font-family
+  --utrecht-alert-heading-font-size
+  --utrecht-alert-heading-font-weight
+
+  Reworked the Alert react component to allow users to add content like heading and paragraph in the children instead of duplicating heading props in the alert directly.
+  Added icon overwrite support in case you want a different icon on a specific alert type.
+
+### Patch Changes
+
+- d2c235c: Uitloggen icoon toegevoegd
+- Updated dependencies [8e508f9]
+- Updated dependencies [29bfdd5]
+- Updated dependencies [965c2a4]
+  - @rijkshuisstijl-community/components-css@6.0.0
+
 ## 1.3.9
 
 ### Patch Changes